requirements.txt
fastapi
uvicorn
librosa
acoustics
numpy
pandas
scipy
pyfunctional
openpyxl
python-multipart
websockets
nptdms
streamlit
plotly
requests
soundfile